fix: minor fixes in scripts and docs

+++ Makefile
@@ -7,10 +7,8 @@
 PRECOMMIT_VERSION := 3.7.1
 
 env-helper:
-	# $(MAKE) -C projects/env_helper build-image run-container HOST_ENV_PATH="$(PWD)/.env-test"
-
 	mkdir -p out/
-	cp .env out/.env
+	if [ -f .env ]; then cp .env out/.env; fi
 	docker build -t env_helper projects/env_helper
 	docker run --rm -it -v "$(PWD)/out":/tmp/out env_helper
 	cp out/.env .env
